Create file for forms and relocate all forms there
