+++ Server/GGuide/models.py
@@ -1,6 +1,4 @@
 from django.db import models
-from django import forms
-from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 from django.contrib.auth import get_user_model
 from django.urls import reverse
@@ -34,27 +32,6 @@ class Article(models.Model):
         return reverse('detail', args=[self.slug])
 
 
-class ArticleForm(forms.Form):
-    form_article_image = forms.ImageField()
-
-
-class SignUpForm(UserCreationForm):
-    username = forms.CharField(max_length=10)
-    email = forms.CharField(max_length=22)
-
-    class Meta:
-        model = User
-        fields = ('username', 'email')
-
-    def __str__(self):
-        return f'{self.username}'
-
-
-class Userlogin(forms.Form):
-    username = forms.CharField(label='username', max_length=10)
-    password = forms.CharField(label='password', max_length=22)
-
-
 class ProfileModel(models.Model):
     img = models.ImageField()
     user = models.OneToOneField(User, on_delete=models.CASCADE)
@@ -62,19 +39,6 @@ class ProfileModel(models.Model):
 
     def __str__(self):
         return f'{self.user}'
-
-
-class ProfileForm(forms.Form):
-    Image = forms.ImageField()
-
-
-class FriendForm(forms.Form):
-    username = forms.CharField(label='username', max_length=10)
-    email = forms.CharField(label='email', max_length=22)
-
-
-class CommentsForm(forms.Form):
-    comment = forms.CharField(label='comment', max_length=250)
 
 
 class Comments(models.Model):
